be/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.1
django-model-utils==5.0.0
django-notifications-hq==1.8.3
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.10==3.1.0
MarkupSaf
itypes==1.2.0
Jinja2==3.1.4
jsonfielde==2.1.5
openapi-codec==1.3.2
packaging==24.1
psycopg2==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
setuptools==75.1.0
simplejson==3.19.3
sqlparse==0.5.1
swapper==1.4.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wheel==0.44.0
